perf: improve calltree render performance
Startup time (to render timeline) is dramtically increased by only rendering deltas on the calltree.

**On click from timeline**
Will render only the part of the stack involved in the element that was clicked.

**Expand single**
Will render only a single level below on demand.

**Expand all**
Will render anything that has not already been rendered.
This means it is slower than it was before but the timeline is faster to render instead.
We are working on the assumption that expand all is less frequently performed.