feat(timeline): add dragging to time line
The timeline will now take up 75% of the height meaning tall logs can go outside
the visible view of the canvas.

To ensure the node that are outside the view can be clicked, you can drag the
canvas to bring them onto the view.

Any canvas can be dragged in the x direction.
Any canvas can be dragged in the Y
direction, will only allow for the tallest
node to be dragged 1/4 from the top of the timeline.