fix: load apex log for analysis command
The comand was not working as createView was being called with the log id
instead of the full log path.

The full log path is needed at the folder to the log is added as a resource root
 when createView is called.

This allows the fetch command to succesfully access files in the folder.