fix: total duration reported incorrrectly sometimes
If EXECUTION_FINISHED was not the last valid line in the log
then the total duration on the status woud have the end time of the valid line instead of the
EXECUTION_FINISHED line.

Similarly the width of the log would be rendered based on the end time of the valid line instead
of EXECUTION_FINISHED + we would see a blank space at the end of the timeline.