fix: timeline not showing all events with a duration
If the events occured outside the `EXECUTION_STARTED` + `EXECUTION_FINISHED`
events they would not be shown.

We now use the last event with a duration and exitStamp for the exittime / duration.
The reason we check for a duration and do not just take the last event
is that an event without a duration is not going to be shown on the timeline
and in all likely hood is simply infomation and does not contribute towards
execution time.