fix: some rows not expanding on calltree
This was due to the endtime being used a the last log element with a duration.
Now there are two properties on root for time
`executionEndTime` which is the endtime of the last node with a duration
(the latest time that should be shown on timeline)
`exitStamp` which is the end time of the lastest node regardless of duration.

This means the total time of the log will sometimes be different to the timeline
time if there are log lines that show extra info e.g `CUMULATIVE_USAGE`