fix: vf_apex_call_start for ApexPage message
They would cause the log to be displayed incorrectly due to any
vf_apex_call_start event for ApexPage message not having a
corresponding vf_apex_call_end.

If a vf_apex_call_start for ApexPage message is detected it will have the
exitTypes for that instance set to an empty array.
This will mean it is ignored during parsing and will become a detail
of the parent instead of method with a start and finish.