refactor: update open log command error message
Opening the log analysis from command
palette may fail if the file is larger than 50mb.
This is because `window.activeTextEditor`
will return undefined for a file over 50mb.
Undefined is returned if there is no file or is too large.

The message will tell them it is too big and to use
either the editor or explorer commands instead.