feat: open file ask for workspace when > 1 match
Only ask for the workspace if there is more than one matching workspace.
This means more than one workspace can be added without the user being asked to select a workspace when opening a file.

The user will only be asked to select a workspace if there is more than once workspace in which a matching file was found.
e.g I can have 10 workspaces added, 2 have a matching file,
the user will only be asked to select from one of the two
matching workspaces.