fix: analysis total time sum unique stacks
Collect the stacks and only sum the unique entries e.g
If we have an analysis like this

methodA: 5s
methodB: 3s
methodC: 2s

Sibling Methods
The three methods are at the same level directly called from EXECUTION_STARTED.
All 3 should be summed as they are in different stacks.

Total: 10s

Child Methods

methodB is a child of methodA and methodC is a sibling of methodA.
we first add methodA to the total but the total for methodB has already been accounted by including methodA so we skip it and any other children of methodA.
methodC is in a unique stack and needs to be added to the total.

Total: 7s